AI chatbots as digital legacy - Code/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.44.0
requests==2.32.3
pandas>=2.2.3
httpx<0.28
